STYLE: Move C++ files to /src
