+++ lrst.i
@@ -3,7 +3,9 @@ lrst.i: SWIG module defining the Python wrapper for our C++ modules
 [Ref.: https://www.swig.org/Doc1.3/Python.html]
  */
 
-%module lrst
+%module lrst  // Module name
+
+// Headers
 %{
 #include <string>
 #include <vector>
@@ -24,7 +26,7 @@ namespace std{
   %template(Int64Vector) vector<int64_t>;
 };
 
-// These are the header functions wrapped by our lrst module
+// These are the header functions wrapped by our lrst module (Like an 'import')
 %include "Python_to_CXX_class.h"  // Contains InputPara for passing parameters to C++
 %include "CXX_to_python_class.h"  // Contains data structures for storing statistics for each file type
 %include "LRST_function.h"  // Functions for calling the C++ statistics computation modules
