+++ src/BAM_module.cpp
