+++ src/BamReader.cpp
