+++ src/BamReader.h
