+++ src/ComStruct.cpp
