+++ src/ComStruct.h
