+++ src/FASTA_module.cpp
