+++ src/kseq.cpp
